audio_extract/requirements.txt
# Audio Extract Module Requirements

# Google Drive API dependencies:
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# Optional: For configuration files
pyyaml>=6.0

# For development and testing:
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Optional: For better logging
colorlog>=6.7.0